requirements.txt
# pip install -r requirements.txt

python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.30.1
gunicorn==22.0.0
pytest==8.2.2
requests~=2.32.3
dataclasses~=0.6
redis~=5.0.5